+++ app/src/main/java/com/example/saar/Contact/ContactFragment.java
@@ -1,4 +1,4 @@
-package com.example.saar;
+package com.example.saar.Contact;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -9,6 +9,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+
+import com.example.saar.R;
 
 public class ContactFragment extends Fragment implements View.OnClickListener {
