+++ app/src/main/java/com/example/saar/Gallery/GalleryFragment.java
@@ -1,4 +1,4 @@
-package com.example.saar;
+package com.example.saar.Gallery;
 
 import android.app.ProgressDialog;
 import android.os.Bundle;
@@ -10,40 +10,48 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
+
+import com.example.saar.Gallery.Gallery;
+import com.example.saar.Gallery.GalleryAdapter;
+import com.example.saar.R;
+import com.example.saar.Retrofit.GetDataService;
+import com.example.saar.Retrofit.RetrofitClientInstance;
+
 import java.util.List;
+
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 
-public class TimelineFragment extends Fragment {
+public class GalleryFragment extends Fragment {
 
     private RecyclerView recyclerView;
-    private EventAdapter mAdapter;
+    private GalleryAdapter mAdapter;
     ProgressDialog progressDialog;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        final View rootView = inflater.inflate(R.layout.fragment_timeline, container, false);
+        final View rootView = inflater.inflate(R.layout.fragment_gallery, container, false);
 
-        recyclerView = rootView.findViewById(R.id.recycler_view_timeline);
+        recyclerView = rootView.findViewById(R.id.recycler_view_gallery);
         progressDialog = new ProgressDialog(rootView.getContext());
         progressDialog.setMessage("Loading....");
         progressDialog.show();
 
         /*Create handle for the RetrofitInstance interface*/
         GetDataService service = RetrofitClientInstance.getRetrofitInstance().create(GetDataService.class);
-        Call<List<Event>> call = service.getAllEvents();
-        call.enqueue(new Callback<List<Event>>() {
+        Call<List<Gallery>> call = service.getAllPhotos();
+        call.enqueue(new Callback<List<Gallery>>() {
             @Override
-            public void onResponse(Call<List<Event>> call, Response<List<Event>> response) {
+            public void onResponse(Call<List<Gallery>> call, Response<List<Gallery>> response) {
                 progressDialog.dismiss();
-                generateDataList(response.body(),rootView);
+                generateDataList(response.body(), rootView);
             }
 
             @Override
-            public void onFailure(Call<List<Event>> call, Throwable t) {
+            public void onFailure(Call<List<Gallery>> call, Throwable t) {
                 progressDialog.dismiss();
                 Toast.makeText(rootView.getContext(), "Please Check Internet Connection", Toast.LENGTH_SHORT).show();
             }
@@ -52,9 +60,9 @@ public class TimelineFragment extends Fragment {
         return rootView;
     }
 
-    private void generateDataList(List<Event> eventList, View rootView) {
-        recyclerView = rootView.findViewById(R.id.recycler_view_timeline);
-        mAdapter = new EventAdapter(eventList);
+    private void generateDataList(List<Gallery> photoList, View rootView) {
+        recyclerView = rootView.findViewById(R.id.recycler_view_gallery);
+        mAdapter = new GalleryAdapter(photoList);
         RecyclerView.LayoutManager layoutManager = new LinearLayoutManager(getContext());
         recyclerView.setLayoutManager(layoutManager);
         recyclerView.setAdapter(mAdapter);
@@ -64,6 +72,6 @@ public class TimelineFragment extends Fragment {
     public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
         //you can set the title for your toolbar here for different fragments different titles
-        getActivity().setTitle(R.string.timeline);
+        getActivity().setTitle(R.string.gallery_fragment);
     }
 }
