+++ app/src/main/java/com/example/saar/Login_SignUp/LoginFragment.java
@@ -1,4 +1,4 @@
-package com.example.saar;
+package com.example.saar.Login_SignUp;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -7,19 +7,22 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class HomeFragment extends Fragment {
+import com.example.saar.R;
+
+public class LoginFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false);
+        return inflater.inflate(R.layout.fragment_login, container, false);
     }
 
     @Override
     public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
         //you can set the title for your toolbar here for different fragments different titles
-        getActivity().setTitle(R.string.app_name);
+        getActivity().setTitle(R.string.saar_login);
     }
+
 }
