+++ app/src/main/java/com/example/saar/Login_SignUp/LoginSignupActivity.java
@@ -1,4 +1,4 @@
-package com.example.saar;
+package com.example.saar.Login_SignUp;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -8,6 +8,8 @@ import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
 import android.view.MenuItem;
+
+import com.example.saar.R;
 
 public class LoginSignupActivity extends AppCompatActivity implements BottomNavigationView.OnNavigationItemSelectedListener {
     @Override
