+++ app/src/main/java/com/example/saar/Login_SignUp/SignUpFragment.java
@@ -1,4 +1,4 @@
-package com.example.saar;
+package com.example.saar.Login_SignUp;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -6,6 +6,8 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.saar.R;
 
 public class SignUpFragment extends Fragment {
