+++ app/src/main/java/com/example/saar/Share/ShareFragment.java
@@ -1,4 +1,4 @@
-package com.example.saar;
+package com.example.saar.Share;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -9,6 +9,9 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageButton;
+
+import com.example.saar.Constant;
+import com.example.saar.R;
 
 public class ShareFragment extends Fragment implements View.OnClickListener {
