+++ app/src/main/java/com/example/saar/Timeline_Events/Event.java
@@ -1,4 +1,4 @@
-package com.example.saar;
+package com.example.saar.Timeline_Events;
 
 import com.google.gson.annotations.SerializedName;
