+++ app/src/main/res/layout/activity_login_signup.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".LoginSignupActivity">
+    tools:context=".Login_SignUp.LoginSignupActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
