+++ app/src/main/res/layout/fragment_about_us.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_margin="@dimen/no_margin"
-    tools:context=".AboutUsFragment">
+    tools:context=".About.AboutUsFragment">
 
     <android.support.v4.widget.NestedScrollView
         android:layout_width="match_parent"
