+++ app/src/main/res/layout/fragment_donate.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DonateFragment"
+    tools:context=".Donate.DonateFragment"
     android:orientation="vertical"
     android:padding="@dimen/padding_medium">
