+++ app/src/main/res/layout/fragment_gallery.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GalleryFragment"
+    tools:context=".Gallery.GalleryFragment"
     android:background="@color/about_us_background">
 
     <android.support.v7.widget.RecyclerView
