+++ app/src/main/res/layout/fragment_sign_up.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignUpFragment">
+    tools:context=".Login_SignUp.SignUpFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
